+++ implementation/src/main/java/com/avioconsulting/mule/testing/transformers/xml/XMLGroovyParserTransformer.groovy
@@ -3,15 +3,15 @@ package com.avioconsulting.mule.testing.transformers.xml
 import com.avioconsulting.mule.testing.muleinterfaces.MuleMessageTransformer
 import com.avioconsulting.mule.testing.muleinterfaces.wrappers.ConnectorInfo
 import com.avioconsulting.mule.testing.muleinterfaces.wrappers.EventWrapper
-import com.avioconsulting.mule.testing.transformers.ClosureMuleMessageHandler
+import com.avioconsulting.mule.testing.transformers.ClosureCurrier
 import groovy.xml.XmlUtil
 
 class XMLGroovyParserTransformer<T extends ConnectorInfo> extends
         XMLTransformer<T> implements
-        MuleMessageTransformer<T>,
-        ClosureMuleMessageHandler {
+        MuleMessageTransformer<T> {
     private final Closure closure
     private final XMLMessageBuilder.MessageType messageType
+    private final ClosureCurrier closureCurrier = new ClosureCurrier<T>()
 
     XMLGroovyParserTransformer(Closure closure,
                                XMLMessageBuilder.MessageType messageType) {
@@ -23,10 +23,10 @@ class XMLGroovyParserTransformer<T extends ConnectorInfo> extends
                            T connectorInfo) {
         def xmlString = connectorInfo.incomingBody ?: muleEvent.messageAsString
         def node = new XmlParser().parseText(xmlString) as Node
-        def forMuleMsg = withMuleEvent(closure,
-                                       muleEvent)
-        def reply = forMuleMsg(node)
-
+        def closure = closureCurrier.curryClosure(this.closure,
+                                                  muleEvent,
+                                                  connectorInfo)
+        def reply = closure(node)
         String outputXmlString
         if (reply instanceof File) {
             outputXmlString = reply.text