+++ implementation/src/main/java/com/avioconsulting/mule/testing/transformers/xml/XMLJAXBTransformer.groovy
@@ -3,17 +3,17 @@ package com.avioconsulting.mule.testing.transformers.xml
 import com.avioconsulting.mule.testing.muleinterfaces.MuleMessageTransformer
 import com.avioconsulting.mule.testing.muleinterfaces.wrappers.ConnectorInfo
 import com.avioconsulting.mule.testing.muleinterfaces.wrappers.EventWrapper
-import com.avioconsulting.mule.testing.transformers.ClosureMuleMessageHandler
+import com.avioconsulting.mule.testing.transformers.ClosureCurrier
 import groovy.util.logging.Log4j2
 
 @Log4j2
 class XMLJAXBTransformer<T extends ConnectorInfo> extends
         XMLTransformer implements
-        MuleMessageTransformer,
-        ClosureMuleMessageHandler {
+        MuleMessageTransformer {
     private final Closure closure
     private final JAXBMarshalHelper helper
     private final XMLMessageBuilder.MessageType messageType
+    private final ClosureCurrier closureCurrier = new ClosureCurrier<T>()
 
     XMLJAXBTransformer(Closure closure,
                        Class inputJaxbClass,
@@ -27,9 +27,10 @@ class XMLJAXBTransformer<T extends ConnectorInfo> extends
                            ConnectorInfo connectorInfo) {
         def strongTypedPayload = helper.unmarshal(event,
                                                   connectorInfo)
-        def forMuleMsg = withMuleEvent(this.closure,
-                                       event)
-        def reply = forMuleMsg(strongTypedPayload)
+        def closure = closureCurrier.curryClosure(this.closure,
+                                                  event,
+                                                  connectorInfo)
+        def reply = closure(strongTypedPayload)
         String xml = reply instanceof File ? reply.text : helper.getMarshalled(reply)
         this.xmlMessageBuilder.build(xml,
                                      event,